move content into component
