+++ embc-app/ClientApp/src/app/admin-dashboard/admin-evacuees/admin-evacuees.component.html
@@ -1,3 +1,22 @@
-<p>
-  admin-evacuees works!
-</p>
+<div class="mb-4">
+  <h1>EMBC Admin Dashboard</h1>
+  <h3>Welcome to your Dashboard!</h3>
+  <p>Here you'll be able to add task numbers, organizations and users, as well as register and maintain evacuees.
+  </p>
+  <h3 class="blue-header">ALL EVACUEES</h3>
+
+  <app-search-bar #searchBar (valueChange)="doSearch($event)"
+    placeholder="Search by task number / last name / registration number / community ...">
+  </app-search-bar>
+
+  <!-- TODO: <a class="small" href="">advanced search</a> -->
+  <!-- TODO: show icons for HOH, RESTRICTED FILE -->
+</div>
+<div>
+  <!-- The start of the table stuff. -->
+  <!-- <p>1 - 10 of 2000 <span class="font-weight-bold">ACTIVE</span> evacuees
+      <small class="ml-3">(<a href="">show deactivated</a>)</small>
+    </p> -->
+  <app-evacuee-search-results #searchResultsView [searchResults]="searchResults$ | async">
+  </app-evacuee-search-results>
+</div>
